requirements.txt
absl-py==0.7.0
aiohttp==3.5.4
aioice==0.6.12
aiortc==0.9.18
aliyun-python-sdk-core==2.13.20
aliyun-python-sdk-core-v3==2.13.10
aliyun-python-sdk-green==3.5.1
aliyun-python-sdk-kms==2.10.1
appdirs==1.4.3
asgiref==3.3.1
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.1.1
Automat==0.6.0
av==6.1.2
awscli==1.18.131
backcall==0.1.0
beautifulsoup4==4.9.3
black==18.9b0
bleach==1.5.0
blinker==1.4
botocore==1.17.54
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloud-init==18.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
configparser==5.0.0
constantly==15.1.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
countdown-timer==0.0.3
crc32c==1.7
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.12
dataclasses==0.8
decorator==4.4.2
deepface==0.0.18
defusedxml==0.6.0
Deprecated==1.2.13
detecto==1.1.6
distlib==0.3.0
distro-info==0.18
Django==3.1.3
django-cors-headers==3.7.0
django-filter==21.1
django-fluent-contents==2.0.7
django-fluent-utils==2.0.1
django-log-request-id==1.6.0
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
django-parler==2.2
django-polymorphic==3.0.0
django-sequences==2.6
django-sslserver==0.22
django-tag-parser==3.2
django-template-analyzer==1.6.2
djangorestframework==3.12.2
djangorestframework-oauth==1.1.0
dlib==19.16.0
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
easydict==1.9
easyfacenet==1.0.5
editdistance==0.5.3
efficientnet==1.0.0
entrypoints==0.3
enum34==1.1.6
essential-generators==0.9.2
face-recognition==1.2.3
face-recognition-models==0.3.0
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.1
Flask-API==1.1
fluentcms-countdown==1.2
fonttools==4.13.0
fpdf==1.7.2
fs==2.4.11
fs.smbfs==0.5.2
future==0.18.2
gast==0.2.2
gdown==3.10.2
ghostscript==0.6
Glances==3.1.4.1
google-cloud==0.34.0
google-images-download==2.5.0
google-pasta==0.1.7
grpcio==1.18.0
h11==0.12.0
h2==4.0.0
h5py==2.10.0
hpack==4.0.0
html5lib==0.9999999
httplib2==0.9.2
hyperframe==6.0.0
hyperlink==17.3.1
idna==2.8
idna-ssl==1.1.0
imageai==2.1.5
imagecleaner==0.4.4
ImageHash==4.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==0.19
importlib-resources==1.2.0
imutils==0.5.2
incremental==16.10.1
inflection==0.5.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.10.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.0
jupyter-console==6.0.0
jupyter-core==4.5.0
jwcrypto==1.0
kaitaistruct==0.9
Keras==2.2.5
Keras-Applications==1.0.8
keras-ocr==0.8.4
Keras-Preprocessing==1.1.0
keras-vggface==0.6
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
lxml==4.5.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
MouseInfo==0.1.3
mtcnn==0.0.9
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.9
networkx==2.4
notebook==5.7.8
num2words==0.5.10
numpy==1.17.0
oauthlib==3.1.1
object-detection==0.1
opencv-contrib-python==4.1.0.25
opencv-python==4.2.0.34
openface==0.2.1
oss2==2.12.1
packaging==19.1
PAM==0.4.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pause==0.3
pbr==5.1.1
pexpect==4.7.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==7.1.2
pluggy==0.12.0
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.12.4
psutil==5.6.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.52
pyclipper==1.2.0
pycocotools==2.0.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pyee==5.0.0
pyexiv2==2.3.1
PyGetWindow==0.0.9
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.5.3
pylibsrtp==0.6.0
pylovepdf==1.3.1
pymongo==3.11.3
PyMsgBox==1.0.9
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.8.1
pypng==0.0.20
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.4
pysmb==1.1.28
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
pytesseract==0.3.4
pytest==5.1.1
python-apt==1.6.2
python-dateutil==2.8.0
python-debian==0.1.32
python-resize-image==1.1.19
python-telegram-bot==12.6.1
python3-xlib==0.15
pyttsx3==2.90
PyTweening==1.0.3
pytz==2019.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.1
reportlab==3.5.23
requests==2.24.0
requests-unixsocket==0.1.5
rsa==4.5
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
samba==0.0.0
scikit-image==0.17.2
scikit-learn==0.20.2
scipy==1.5.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
service-identity==16.0.0
setuptools-rust==0.11.6
Shapely==1.7.0
shortuuid==1.0.8
simplegeneric==0.8.1
simplejson==3.17.0
six==1.12.0
soupsieve==2.0.1
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
telebot==0.0.3
telegram==0.0.1
telegram-send==0.25
tensorboard==1.14.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-slim==1.1.0
tifffile==2020.7.17
toml==0.10.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.2
Twisted==17.9.0
txaio==18.8.1
typing-extensions==3.7.2
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.24.1
utils==0.9.0
uWSGI==2.0.18
validators==0.16.0
virtualenv==20.0.7
wcwidth==0.1.7
webencodings==0.5
webrtc-object-detection==0.1
Werkzeug==0.15.5
widgetsnbextension==3.5.0
wrapt==1.11.2
wsproto==1.0.0
xkit==0.0.0
yarl==1.3.0
zipp==0.6.0
zope.interface==4.3.2
